+++ src/globals.cpp
@@ -4,7 +4,7 @@ int Global::TotalFromChunkX, Global::TotalFromChunkZ, Global::TotalToChunkX, Glo
 int Global::FromChunkX = UNDEFINED, Global::FromChunkZ = UNDEFINED, Global::ToChunkX = UNDEFINED, Global::ToChunkZ = UNDEFINED;
 size_t Global::MapsizeZ = 0, Global::MapsizeX = 0, Global::Terrainsize = 0;
 int Global::MapminY = 0;
-unsigned int Global::MapsizeY = 256;
+size_t Global::MapsizeY = 256;
 int Global::OffsetY = 2;
 Settings Global::settings = { East, false, false, false, false, 0, false, false, false, false };
